Added cleanup trap on exit, improved messaging
There is now a trap on script exit that detects if the script exited due
to the effects of `set -e` and prints a failure message.  This trap is
also where the temp directory gets deleted -- failures meant the
directory stayed around due to premature exit

TODO: If behaviour is wanted so that the temp directory does stick around
after failure, an option to the all.sh script can be added to just print
out where to find the directory.